tests/requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.9.1
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
pydantic==1.8.2
redis==4.1.1
python-dotenv==0.19.2
aiofiles==0.8.0